Package refactoring
