+++ app/src/main/AndroidManifest.xml
@@ -14,9 +14,9 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         tools:ignore="GoogleAppIndexingWarning">
-        <activity android:name=".login.LoginActivity" />
+        <activity android:name="com.bluecat.githubfeed.ui.activities.login.LoginActivity" />
         <activity
-            android:name=".splash.SplashActivity"
+            android:name="com.bluecat.githubfeed.ui.activities.splash.SplashActivity"
             android:screenOrientation="portrait">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -24,6 +24,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".main.MainActivity" />
+        <activity android:name="com.bluecat.githubfeed.ui.activities.main.MainActivity" />
     </application>
 </manifest>
