--- app/src/main/java/com/bluecat/githubfeed/login/LoginActivityView.kt
@@ -1,8 +1,0 @@
-package com.bluecat.githubfeed.login
-
-import com.bluecat.core.BaseView
-
-interface LoginActivityView : BaseView {
-    fun onLoginSuccess(name: String?)
-    fun onLoginFailure(state: String?, needOTP: Boolean)
-}
