+++ app/src/main/java/com/bluecat/githubfeed/presenters/LoginPresenter.kt
@@ -1,10 +1,27 @@
-package com.bluecat.githubfeed.login
+/*
+ * Copyright 2019 Team Mulro in BlueCat-Community
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.bluecat.githubfeed.presenters
 
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.Observer
 import com.bluecat.core.BasePresenter
 import com.bluecat.githubfeed.api.NetworkModule
 import com.bluecat.githubfeed.util.AuthUtil
+import com.bluecat.githubfeed.viewTypes.LoginActivityView
 import timber.log.Timber
 
 class LoginPresenter : BasePresenter<LoginActivityView>() {
