+++ app/src/main/java/com/bluecat/githubfeed/ui/activities/login/LoginActivity.kt
@@ -1,4 +1,20 @@
-package com.bluecat.githubfeed.login
+/*
+ * Copyright 2019 Team Mulro in BlueCat-Community
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.bluecat.githubfeed.ui.activities.login
 
 import android.os.Bundle
 import android.view.View
@@ -6,10 +22,13 @@ import android.widget.Toast
 import com.bluecat.core.BaseActivity
 import com.bluecat.core.qualifiers.RequirePresenter
 import com.bluecat.githubfeed.R
+import com.bluecat.githubfeed.presenters.LoginPresenter
+import com.bluecat.githubfeed.viewTypes.LoginActivityView
 import kotlinx.android.synthetic.main.activity_login.*
 
 @RequirePresenter(LoginPresenter::class)
-class LoginActivity : BaseActivity<LoginPresenter, LoginActivityView>(), LoginActivityView {
+class LoginActivity : BaseActivity<LoginPresenter, LoginActivityView>(),
+    LoginActivityView {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
