+++ app/src/main/java/com/bluecat/githubfeed/ui/adapters/MainPagerAdapter.kt
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2019 Team Mulro in BlueCat-Community
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.bluecat.githubfeed.ui.adapters
+
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentManager
+import androidx.fragment.app.FragmentPagerAdapter
+import com.bluecat.githubfeed.ui.fragments.FeedFragment
+
+class MainPagerAdapter(fm : FragmentManager):FragmentPagerAdapter(fm, BEHAVIOR_RESUME_ONLY_CURRENT_FRAGMENT) {
+
+    override fun getItem(position: Int): Fragment {
+        return when(position)
+        {
+            0-> FeedFragment()
+            else -> FeedFragment()
+        }
+    }
+
+    override fun getCount() = 1
+}
