+++ app/src/main/java/com/bluecat/githubfeed/ui/fragments/FeedFragment.kt
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-package com.bluecat.githubfeed.main
+package com.bluecat.githubfeed.ui.fragments
 
-import com.bluecat.core.BaseView
+import androidx.fragment.app.Fragment
 
-interface MainActivityView : BaseView {
-    fun getGitHubUserInfo(username: String)
+class FeedFragment:Fragment() {
+    //TODO
 }
