+++ app/src/main/java/com/bluecat/githubfeed/viewTypes/SplashActivityView.kt
@@ -14,16 +14,10 @@
  * limitations under the License.
  */
 
-package com.bluecat.githubfeed.splash
+package com.bluecat.githubfeed.viewTypes
 
-import android.os.Handler
-import android.os.Looper
-import com.bluecat.core.BasePresenter
-import timber.log.Timber
+import com.bluecat.core.BaseView
 
-class SplashPresenter : BasePresenter<SplashActivityView>() {
-    init {
-        Timber.d("Initialize SplashPresenter.")
-        Handler(Looper.getMainLooper()).postDelayed({ baseView.moveMain() }, 1000)
-    }
+interface SplashActivityView : BaseView {
+    fun moveMain()
 }
