+++ app/src/main/res/layout/activity_login.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:gravity="center"
     android:orientation="vertical"
-    tools:context=".login.LoginActivity">
+    tools:context=".ui.activities.login.LoginActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
