+++ app/src/main/res/menu/menu_bottom.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item
+        android:id="@+id/navigation_menu1"
+        android:title="Feed" />
+
+    <item
+        android:id="@+id/navigation_menu2"
+        android:title="MENU2" />
+
+    <item
+        android:id="@+id/navigation_menu3"
+        android:title="MENU3" />
+
+    <item
+        android:id="@+id/navigation_menu4"
+        android:title="MENU3" />
+
+</menu>
